requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
transformers
torch
pillow
pytest
dotenv
requests
python-multipart